// src/app/tabs/shared/models/consumption-alert.model.ts
export class ConsumptionAlert {
    id?: number;
    fromBillPeriodId?: number;
    fromBillPeriod?: string;
    toBillPeriodId?: number;
    toBillPeriod?: string;
    utilityTypeId?: number;
    utilityType?: string;
    consumptionType?: string;
    consumptionTypeId?: number;
    description?: string;
    percentage?: number;
}